Remove legacy generate and bundle cmds
Removed legacy "generate csv/packagemanifests/bundle", and "bundle create/validate" commands.
Also removed associated scaffolding packages for these commands in "internal/generate/olm-catalog".
