+++ changelog/fragments/rm-legacy-bundle-cmds.yaml
@@ -1,0 +1,11 @@
+# entries is a list of entries to include in
+# release notes and/or the migration guide
+entries:
+  - description: >
+      Remove legacy "bundle create/validate" commands
+    kind: "removal"
+    # Is this a breaking change?
+    breaking: true
+    migration:
+      header: Remove legacy "bundle create/validate" commands
+      body: TBD
