+++ changelog/fragments/rm-legacy-gen-csv.yaml
@@ -1,0 +1,11 @@
+# entries is a list of entries to include in
+# release notes and/or the migration guide
+entries:
+  - description: >
+      Remove legacy generate csv command
+    kind: "removal"
+    # Is this a breaking change?
+    breaking: true
+    migration:
+      header: Remove legacy generate csv command
+      body: TBD
