+++ cmd/operator-sdk/cli/legacy.go
@@ -21,9 +21,7 @@ import (
 
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/add"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/build"
-	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/bundle"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/completion"
-	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/generate"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/new"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/olm"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/scorecard"
@@ -65,10 +63,8 @@ func GetCLIRoot() *cobra.Command {
 	root.AddCommand(
 		add.NewCmd(),
 		build.NewCmd(),
-		bundle.NewCmdLegacy(),
 		scorecard.NewCmd(),
 		completion.NewCmd(),
-		generate.NewCmdLegacy(),
 		new.NewCmd(),
 		olm.NewCmd(),
 		version.NewCmd(),
