+++ website/content/en/docs/cli/operator-sdk.md
@@ -19,9 +19,7 @@ An SDK for building operators with ease
 
 * [operator-sdk add](../operator-sdk_add)	 - Adds a controller or resource to the project
 * [operator-sdk build](../operator-sdk_build)	 - Compiles code and builds artifacts
-* [operator-sdk bundle](../operator-sdk_bundle)	 - Manage operator bundle metadata
 * [operator-sdk completion](../operator-sdk_completion)	 - Generators for shell completions
-* [operator-sdk generate](../operator-sdk_generate)	 - Invokes a specific generator
 * [operator-sdk new](../operator-sdk_new)	 - Creates a new operator application
 * [operator-sdk olm](../operator-sdk_olm)	 - Manage the Operator Lifecycle Manager installation in your cluster
 * [operator-sdk scorecard](../operator-sdk_scorecard)	 - Runs scorecard
