--- website/content/en/docs/cli/operator-sdk_bundle.md
@@ -1,31 +1,0 @@
----
-title: "operator-sdk bundle"
----
-## operator-sdk bundle
-
-Manage operator bundle metadata
-
-### Synopsis
-
-Manage bundle builds, bundle metadata generation, and bundle validation.
-An operator bundle is a portable operator packaging format understood by Kubernetes
-native software, like the Operator Lifecycle Manager.
-
-More information about operator bundles and metadata:
-https://github.com/operator-framework/operator-registry/blob/master/docs/design/operator-bundle.md
-
-More information about the integration with OLM via SDK:
-https://sdk.operatorframework.io/docs/olm-integration/legacy
-
-
-### Options
-
-```
-  -h, --help   help for bundle
-```
-
-### SEE ALSO
-
-* [operator-sdk](../operator-sdk)	 - An SDK for building operators with ease
-* [operator-sdk bundle validate](../operator-sdk_bundle_validate)	 - Validate an operator bundle
-
