--- website/content/en/docs/cli/operator-sdk_generate.md
@@ -1,24 +1,0 @@
----
-title: "operator-sdk generate"
----
-## operator-sdk generate
-
-Invokes a specific generator
-
-### Synopsis
-
-The 'operator-sdk generate' command invokes a specific generator to generate
-code or manifests.
-
-### Options
-
-```
-  -h, --help   help for generate
-```
-
-### SEE ALSO
-
-* [operator-sdk](../operator-sdk)	 - An SDK for building operators with ease
-* [operator-sdk generate bundle](../operator-sdk_generate_bundle)	 - Generates bundle data for the operator
-* [operator-sdk generate packagemanifests](../operator-sdk_generate_packagemanifests)	 - Generates a package manifests format
-
